+++ packages/account-api/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Uncategorized
+
+- feat(account-api): add `isBip44Account` helper ([#319](https://github.com/MetaMask/accounts/pull/319))
+
 ### Added
 
 - Add `MultichainAccountWallet.sync` method ([#321](https://github.com/MetaMask/accounts/pull/321))